refactor(main): reubicacion de archivos
