+++ backend/models/asistentes.py
@@ -1,14 +1,18 @@
 from typing import Optional
 from pydantic import BaseModel, EmailStr
+from nanoid import generate
 
 
 class AsistenteBase(BaseModel):
+    clave: str = ""
     folio: str
     primer_nombre: str
     segundo_nombre: Optional[str] = None
     apellido_p: str
     apellido_m: str
     correo: EmailStr
+    ya_descargo: bool = False
+    evento: str = ""
 
     @property
     def nombre_completo(self) -> str:
@@ -34,7 +38,7 @@ class AsistenteCreate(BaseModel):
 
 class AsistenteUpdate(BaseModel):
     primer_nombre: Optional[str] = None
-    segundo_nombre: Optional[str] = None
+    segundo_nAnyHttpUrlombre: Optional[str] = None
     apellido_p: Optional[str] = None
     apellido_m: Optional[str] = None
     correo: Optional[EmailStr] = None
